rest/requirements_new.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.19.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.2.1
python-dotenv==1.0.1
jinja2==3.1.2
pydantic==2.10.3
pydantic-settings==2.0.3
alembic==1.14.0